+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added color diff for snapshot testing
+- Added Bool Type, `true` and `false` literals
 
 ### Changed
 ### Fixed